+++ backend-rust/CHANGELOG.md
@@ -8,7 +8,9 @@ Database schema version: 5
 
 ### Fixed
 
-- Add database migration fixing invalid delegator state, caused by validator/baker getting removed or changing status to 'ClosedForAll' without moving delegators to the passive pool.
+- Add database migration fixing:
+  - Invalid bakers caused by `DelegationEvent::RemoveBaker` event not being handled by the indexer until now.
+  - Invalid delegator state, caused by validator/baker getting removed or changing status to 'ClosedForAll' without moving delegators to the passive pool.
 - Fixed indexer missing handling of moving delegators as pool got removed or closed.
 - Fixed indexer missing handling of event of baker switching directly to delegation.